# cmd/docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "contact": {
      "email": "bangadam.dev@gmail.com",
      "name": "Email Developer"
    },
    "description": "API documentaion of Backend Test Majoo",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService": "http://swagger.io/terms/",
    "title": "API documentaion of Backend Test Majoo",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "description": "Do Login User",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/domain.LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response.AppSuccess"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response.AppError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response.AppError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response.AppError"
                }
              }
            }
          }
        }
      }
    },
    "/report-transaction-merchant": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Report Transaction"
        ],
        "parameters": [
          {
            "description": "page filter",
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "data limit filter",
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "start date filter",
            "in": "query",
            "name": "start_date",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "end date filter",
            "in": "query",
            "name": "end_date",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response.AppSuccess"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response.AppError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response.AppError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response.AppError"
                }
              }
            }
          }
        }
      }
    },
    "/report-transaction-outlet/:outlet_id": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Report Transaction"
        ],
        "parameters": [
          {
            "description": "page filter",
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "data limit filter",
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "start date filter",
            "in": "query",
            "name": "start_date",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "end date filter",
            "in": "query",
            "name": "end_date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "outlet_id",
            "description": "outlet id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response.AppSuccess"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response.AppError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response.AppError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response.AppError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "domain.LoginRequest": {
        "properties": {
          "password": {
            "example": "admin1",
            "type": "string"
          },
          "username": {
            "example": "admin1",
            "type": "string"
          }
        },
        "type": "object"
      },
      "response.AppError": {
        "properties": {
          "message": {
            "description": "Message is the error message that may be displayed to end users",
            "type": "string"
          },
          "meta": {
            "description": "Meta is the error detail detail data"
          }
        },
        "type": "object"
      },
      "response.AppSuccess": {
        "properties": {
          "data": {},
          "meta": {}
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}